Move all Espresso helpers into a single helper file
